pyproject.toml
[tool.poetry]
name = "llmeval"
version = "0.1.0"
description = "An open-source LLM evaluation and regression testing framework"
authors = ["Emily Suh <emilysuh.dev@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/emilysuh/llm-eval-framework"
repository = "https://github.com/emilysuh/llm-eval-framework"
keywords = ["llm", "evaluation", "testing", "regression", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
packages = [{ include = "llmeval" }]

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "^0.40.0"
openai = "^1.50.0"
pydantic = "^2.9.0"
typer = { version = "^0.12.0", extras = ["all"] }
rich = "^13.9.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
ruff = "^0.7.0"
black = "^24.10.0"
mypy = "^1.13.0"
types-pyyaml = "^6.0.12"

[tool.poetry.scripts]
llmeval = "llmeval.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=llmeval --cov-report=term-missing --cov-fail-under=80"
# Pydantic domain models (TestCase, TestSuite, TestResult) share pytest's
# "Test*" collection prefix but are not test classes.
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]

[tool.coverage.run]
omit = ["tests/*"]
